@@ -1,33 +1,31 @@
 {
   "dependencies": {
-    "@astrojs/mdx": "^3.1.9",
+    "@astrojs/mdx": "^4.0.8",
     "@astrojs/rss": "^4.0.11",
-    "@astrojs/sitemap": "^1.4.0",
-    "astro": "^4.16.18",
+    "@astrojs/sitemap": "^3.2.1",
+    "astro": "^5.2.4",
     "bun": "^1.2.2",
-    "modern-normalize": "^2.0.0",
+    "modern-normalize": "^3.0.1",
     "posthog-js": "^1.215.3",
-    "sharp": "^0.32.6",
+    "sharp": "^0.33.5",
     "thumbhash": "^0.1.1"
   },
   "devDependencies": {
-    "@astrojs/check": "^0.7.0",
+    "@astrojs/check": "^0.9.4",
     "@types/bun": "^1.2.2",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
-    "eslint": "^8.57.1",
-    "eslint-config-prettier": "^9.1.0",
+    "@types/eslint-config-prettier": "^6.11.3",
+    "eslint": "^9.19.0",
+    "eslint-config-prettier": "^10.0.1",
     "eslint-plugin-astro": "^1.3.1",
     "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-prettier": "^5.2.3",
-    "postcss-preset-env": "^8.5.1",
     "prettier": "^3.4.2",
-    "prettier-plugin-astro": "^0.13.0",
-    "typescript": "^5.7.3"
+    "prettier-plugin-astro": "^0.14.1",
+    "typescript": "^5.7.3",
+    "typescript-eslint": "^8.23.0"
   },
   "name": "blog",
   "scripts": {
-    "astro": "astro",
     "build": "astro build",
     "check": "astro check",
     "dev": "astro dev",
