@@ -1,7 +1,7 @@
 // @ts-check
 
 /** @type {import('prettier').Config} */
-module.exports = {
+export default {
   semi: true,
   singleQuote: true,
   plugins: ['prettier-plugin-astro'],
